# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Character Service API
    description: TODO character API
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /character/{characterId}/inventory/{id}/item:
        get:
            tags:
                - inventory
            summary: showItem inventory
            description: Show items in an inventory
            operationId: inventory#showItem
            parameters:
                - name: view
                  in: query
                  description: View to render
                  required: false
                  type: string
                  enum:
                    - default
                    - tiny
                - name: characterId
                  in: path
                  description: ID of character to show
                  required: true
                  type: string
                - name: id
                  in: path
                  description: ID of inventory to show
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/InventoryStoredItemResponseTinyCollection'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/InventoryShowItemNotFoundResponseBody'
                        required:
                            - message
                            - id
            schemes:
                - http
    /character/{characterId}/inventory/{id}/item/{itemId}:
        post:
            tags:
                - inventory
            summary: addItem inventory
            description: Add new item to inventory.
            operationId: inventory#addItem
            parameters:
                - name: characterId
                  in: path
                  description: ID of character to show
                  required: true
                  type: string
                - name: id
                  in: path
                  description: ID of inventory to add
                  required: true
                  type: string
                - name: itemId
                  in: path
                  description: ID of item to add
                  required: true
                  type: string
                - name: AddItemRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/InventoryAddItemRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/InventoryAddItemResponseBody'
            schemes:
                - http
        delete:
            tags:
                - inventory
            summary: removeItem inventory
            description: Remove an item from inventory
            operationId: inventory#removeItem
            parameters:
                - name: characterId
                  in: path
                  description: ID of character to show
                  required: true
                  type: string
                - name: id
                  in: path
                  description: ID of inventory to add
                  required: true
                  type: string
                - name: itemId
                  in: path
                  description: ID of item to add
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/InventoryRemoveItemResponseBody'
            schemes:
                - http
...
definitions:
...
    InventoryAddItemRequestBody:
        title: InventoryAddItemRequestBody
        type: object
        properties:
            view:
                type: string
                description: View to render
                example: tiny
                enum:
                    - default
                    - tiny
        example:
            view: tiny
    InventoryAddItemResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.example.inventory; view=default'
        type: object
        properties:
            characterId:
                type: string
                description: CharacterId
                example: 123abc
            id:
                type: string
                description: ID is the unique id of the inventory.
                example: 123abc
            items:
                type: array
                items:
                    $ref: '#/definitions/StoredItemResponseBodyTiny'
                description: Character items
                example:
                    - damage: 37.8267
                      healing: 12.6
                      id: 123abc
                      name: Boots of travel
                      protection: 65.21
                    - damage: 37.8267
                      healing: 12.6
                      id: 123abc
                      name: Boots of travel
                      protection: 65.21
                    - damage: 37.8267
                      healing: 12.6
                      id: 123abc
                      name: Boots of travel
                      protection: 65.21
        description: AddItemResponseBody result type (default view)
        example:
            characterId: 123abc
            id: 123abc
            items:
                - damage: 37.8267
                  healing: 12.6
                  id: 123abc
                  name: Boots of travel
                  protection: 65.21
                - damage: 37.8267
                  healing: 12.6
                  id: 123abc
                  name: Boots of travel
                  protection: 65.21
        required:
            - id
            - characterId
            - items
    InventoryRemoveItemResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.example.inventory; view=default'
        type: object
        properties:
            characterId:
                type: string
                description: CharacterId
                example: 123abc
            id:
                type: string
                description: ID is the unique id of the inventory.
                example: 123abc
            items:
                type: array
                items:
                    $ref: '#/definitions/StoredItemResponseBodyTiny'
                description: Character items
                example:
                    - damage: 37.8267
                      healing: 12.6
                      id: 123abc
                      name: Boots of travel
                      protection: 65.21
                    - damage: 37.8267
                      healing: 12.6
                      id: 123abc
                      name: Boots of travel
                      protection: 65.21
                    - damage: 37.8267
                      healing: 12.6
                      id: 123abc
                      name: Boots of travel
                      protection: 65.21
        description: RemoveItemResponseBody result type (default view)
        example:
            characterId: 123abc
            id: 123abc
            items:
                - damage: 37.8267
                  healing: 12.6
                  id: 123abc
                  name: Boots of travel
                  protection: 65.21
                - damage: 37.8267
                  healing: 12.6
                  id: 123abc
                  name: Boots of travel
                  protection: 65.21
        required:
            - id
            - characterId
            - items
    InventoryShowItemNotFoundResponseBody:
        title: InventoryShowItemNotFoundResponseBody
        type: object
        properties:
            id:
                type: string
                description: ID of missing resource
                example: Aspernatur exercitationem officiis.
            message:
                type: string
                description: Message of error
                example: resource 1 not found
        description: Character not found
        example:
            id: Vel itaque animi est.
            message: resource 1 not found
        required:
            - message
            - id
...
    InventoryStoredItemResponseTinyCollection:
        title: 'Mediatype identifier: application/json; type=collection; view=tiny'
        type: array
        items:
            $ref: '#/definitions/StoredItemResponseTiny'
        description: InventoryStoredItemResponseTinyCollection is the result type for an array of StoredItemResponseTiny (default view)
        example:
            - damage: 37.8267
              description: Boots of Travel is an item purchasable at the Base Shop, under Accessories. It can be upgraded by purchasing the recipe again.
              healing: 12.6
              id: 123abc
              name: Boots of travel
              protection: 65.21
            - damage: 37.8267
              description: Boots of Travel is an item purchasable at the Base Shop, under Accessories. It can be upgraded by purchasing the recipe again.
              healing: 12.6
              id: 123abc
              name: Boots of travel
              protection: 65.21
            - damage: 37.8267
              description: Boots of Travel is an item purchasable at the Base Shop, under Accessories. It can be upgraded by purchasing the recipe again.
              healing: 12.6
              id: 123abc
              name: Boots of travel
              protection: 65.21
...
    StoredItemResponseBodyTiny:
        title: 'Mediatype identifier: application/json; view=default'
        type: object
        properties:
            damage:
                type: number
                description: Damage
                default: 0
                example: 37.8267
                format: double
            healing:
                type: number
                description: Healing
                default: 0
                example: 12.6
                format: double
            id:
                type: string
                description: ID is the unique id of the Item.
                example: 123abc
            name:
                type: string
                description: Name
                example: Boots of travel
            protection:
                type: number
                description: Protection
                default: 0
                example: 65.21
                format: double
        description: A StoredItem describes a Item retrieved by the db. (tiny view) (default view)
        example:
            damage: 37.8267
            healing: 12.6
            id: 123abc
            name: Boots of travel
            protection: 65.21
        required:
            - id
            - name
    StoredItemResponseTiny:
        title: 'Mediatype identifier: application/json; view=default'
        type: object
        properties:
            damage:
                type: number
                description: Damage
                default: 0
                example: 37.8267
                format: double
            healing:
                type: number
                description: Healing
                default: 0
                example: 12.6
                format: double
            id:
                type: string
                description: ID is the unique id of the Item.
                example: 123abc
            name:
                type: string
                description: Name
                example: Boots of travel
            protection:
                type: number
                description: Protection
                default: 0
                example: 65.21
                format: double
        description: A StoredItem describes a Item retrieved by the db. (tiny view) (default view)
        example:
            damage: 37.8267
            healing: 12.6
            id: 123abc
            name: Boots of travel
            protection: 65.21
        required:
            - id
            - name